Задача на символы в словах

+++ junior/001_Collections_Pro/src/main/java/map/User.java
@@ -35,7 +35,7 @@ public class User {
     public int hashCode() {
         int result = 1;
         result = 31 * result + name.hashCode();
-        result = 31 * result + (int)children;
+        result = 31 * result + children;
         result = 31 * result + birthday.hashCode();
         return result;
     }
